chore:crypto-bundle/CBDL#2.Changed code for linter rules

+++ pkg/nats/common.go
@@ -65,7 +65,7 @@ type consumerConfig interface {
 	GetSubjectName() string
 
 	IsAutoReSubscribeEnabled() bool
-	GetAutoResubscribeCount() uint16
+	GetAutoResubscribeCount() int
 	GetAutoResubscribeDelay() time.Duration
 
 	GetNakDelayTimings() []time.Duration
